pyproject.toml
[project]
name = "mini-vllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "huggingface-hub>=0.36.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.2",
    "tokenizers>=0.22.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
mini-vllm = "mini_vllm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
